MBS-10666 mark legacy gradle testing convention
