+++ build-logic/kotlin-convention/src/main/kotlin/convention.kotlin-jvm.gradle.kts
@@ -1,5 +1,4 @@
 plugins {
     id("kotlin")
     id("convention.kotlin-base")
-    id("convention.gradle-testing")
 }
