+++ libraries/src/main/kotlin/com/avito/android/ExternalLibrariesExtension.kt
@@ -39,6 +39,7 @@ abstract class ExternalLibrariesExtension @Inject constructor(private val provid
 
     val kotlinXCli = "org.jetbrains.kotlinx:kotlinx-cli:0.2.1"
     val kotlinStdlib = "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
+    val kotlinStdlibJdk7 = "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     val kotlinReflect = "org.jetbrains.kotlin:kotlin-reflect:$kotlinVersion"
     val kotlinHtml = "org.jetbrains.kotlinx:kotlinx-html-jvm:0.7.2"
     val kotlinCompilerEmbeddable = "org.jetbrains.kotlin:kotlin-compiler-embeddable:$kotlinVersion"
