+++ subprojects/gradle/artifactory-app-backup/build.gradle.kts
@@ -2,7 +2,7 @@ plugins {
     id("convention.kotlin-jvm")
     id("convention.publish-gradle-plugin")
     id("convention.libraries")
-    id("convention.gradle-test-kit")
+    id("convention.gradle-testing")
 }
 
 dependencies {
