+++ subprojects/gradle/instrumentation-changed-tests-finder/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     id("convention.kotlin-jvm")
     id("convention.publish-gradle-plugin")
+    id("convention.gradle-testing-legacy")
 }
 
 dependencies {
