+++ subprojects/gradle/instrumentation-tests/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
     id("convention.publish-gradle-plugin")
     id("convention.integration-testing")
     id("convention.libraries")
+    id("convention.gradle-testing-legacy")
 }
 
 dependencies {
