+++ subprojects/gradle/test-summary/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     id("convention.kotlin-jvm")
     id("convention.publish-gradle-plugin")
     id("convention.libraries")
+    id("convention.gradle-testing-legacy")
 }
 
 dependencies {
