+++ subprojects/gradle/trace-event/build.gradle.kts
@@ -6,4 +6,8 @@ plugins {
 
 dependencies {
     implementation(libs.gson)
+
+    testImplementation(libs.kotlinStdlibJdk7) {
+        because("kotlin.io.path experimental; consider replacing")
+    }
 }
