+++ subprojects/gradle/trace-event/src/test/kotlin/com/avito/android/trace/TraceReportTest.kt
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import java.nio.file.Path
-import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.createTempFile
+import kotlin.io.path.ExperimentalPathApi
 
 class TraceReportTest {
